made grid spacing equal
use float for grid division

+++ lib/Chart/Grid.cpp
@@ -31,7 +31,7 @@ SOFTWARE.*
 
 #include "Grid.h"
 
-Grid::Grid(GfxItem* parent, GridOrientation orientation, int division, uint16_t color) :
+Grid::Grid(GfxItem* parent, GridOrientation orientation, float division, uint16_t color) :
 	GfxItem(parent, parent->x0(), parent->y0(), parent->width(), parent->height(), color),
 	_orientation(orientation),
 	_division(division) {
@@ -39,14 +39,14 @@ Grid::Grid(GfxItem* parent, GridOrientation orientation, int division, uint16_t 
 
 void Grid::draw() {
 	if (_orientation == vertical) {
-		int x = _x;
+		float x = _x;
 		while (x < (_x + _w  - 1)) {
 			_gfx.drawFastVLine(x, _y, _h, _color);
 			x += _division;
 		}
 	} else
 		if (_orientation == horizontal) {
-			int y = _y + _h - 1;
+			float y = _y + _h - 1;
 			while (y > (_y  - 1)) {
 				_gfx.drawFastHLine(_x, y, _w, _color);
 				y -= _division;
