+++ lib/Chart/PlotArea.h
@@ -40,8 +40,8 @@ class PlotArea : public GfxItem {
 public:
 	PlotArea(GfxItem& parent, int offsetLeft, int offsetRight, int offsetTop, int offsetBottom, uint16_t color);
 
-	Grid& addGridVertical(int division, uint16_t color);
-	Grid& addGridHorizontal(int division, uint16_t color);
+	Grid& addGridVertical(float division, uint16_t color);
+	Grid& addGridHorizontal(float division, uint16_t color);
 	YScale& addYScale(int offset, float minValue, float maxValue, uint16_t color);
